feat(desktop): add cloudblink management page
Add a dedicated CloudBlink page at the account level with storage usage, plan & trial status, billing portal access and an inline plan picker. Extract the plan UI out of the upgrade dialog into a shared SubscriptionPlans component and the settings layout primitives out of the account page so both pages reuse them.
