+++ apps/desktop/src/hooks/queries/use-space.ts
@@ -1,3 +1,4 @@
+import { getErrorCode } from "@blinkdisk/utils/error";
 import { useAccountId } from "@desktop/hooks/use-account-id";
 import { useQueryKey } from "@desktop/hooks/use-query-key";
 import { trpc } from "@desktop/lib/trpc";
@@ -12,6 +13,8 @@ export function useSpace() {
     queryFn: () => {
       return trpc.cloudblink.space.query();
     },
+    retry: (failureCount, error) =>
+      getErrorCode(error) !== "SPACE_NOT_FOUND" && failureCount < 3,
     refetchOnMount: false,
     enabled: !!isOnlineAccount,
   });
