+++ apps/desktop/src/i18n.ts
@@ -13,6 +13,7 @@ i18n.loadNamespaces([
   "directory",
   "backup",
   "subscription",
+  "cloudblink",
   "update",
   "task",
 ]);
