+++ apps/desktop/src/routes/{-$accountId}/account.tsx
@@ -4,6 +4,11 @@ import { Button } from "@blinkdisk/ui/button";
 import { Tooltip, TooltipContent, TooltipTrigger } from "@blinkdisk/ui/tooltip";
 import { Email } from "@desktop/components/accounts/email";
 import { SignOutDialog } from "@desktop/components/dialogs/sign-out";
+import {
+  SettingsGroup,
+  SettingsPanel,
+  SettingsRow,
+} from "@desktop/components/settings";
 import { useUpdateAccountForm } from "@desktop/hooks/forms/use-update-account-form";
 import { useUpdatePreferencesForm } from "@desktop/hooks/forms/use-update-preferences-form";
 import { useAccountList } from "@desktop/hooks/queries/use-account-list";
@@ -293,80 +298,6 @@ function LocalAccountSection() {
         </div>
       </SettingsRow>
     </SettingsPanel>
-  );
-}
-
-type SettingsGroupProps = {
-  title: string;
-  children: ReactNode;
-};
-
-function SettingsGroup({ title, children }: SettingsGroupProps) {
-  return (
-    <section className="grid gap-4">
-      <h2 className="text-xl font-semibold">{title}</h2>
-      {children}
-    </section>
-  );
-}
-
-type SettingsPanelProps = {
-  children: ReactNode;
-};
-
-function SettingsPanel({ children }: SettingsPanelProps) {
-  return (
-    <section className="border-border bg-card overflow-hidden rounded-xl border">
-      {children}
-    </section>
-  );
-}
-
-type SettingsRowProps = {
-  title?: string;
-  description?: string;
-  titleClassName?: string;
-  fullWidth?: boolean;
-  separated?: boolean;
-  children: ReactNode;
-};
-
-function SettingsRow({
-  title,
-  description,
-  titleClassName,
-  fullWidth,
-  separated,
-  children,
-}: SettingsRowProps) {
-  if (fullWidth) {
-    return (
-      <div className="border-border border-b px-5 py-4 last:border-b-0">
-        {children}
-      </div>
-    );
-  }
-
-  return (
-    <div
-      className={`border-border flex flex-col gap-4 border-b px-5 py-4 last:border-b-0 md:min-h-20 md:flex-row md:items-center md:justify-between ${
-        separated ? "border-t" : ""
-      }`}
-    >
-      <div className="min-w-0">
-        {title ? (
-          <p className={titleClassName || "text-base font-medium"}>{title}</p>
-        ) : null}
-        {description ? (
-          <p className="text-muted-foreground mt-1 max-w-sm text-sm">
-            {description}
-          </p>
-        ) : null}
-      </div>
-      <div className="flex w-full justify-start md:w-auto md:justify-end">
-        {children}
-      </div>
-    </div>
   );
 }
