+++ libs/utils/src/error.ts
@@ -37,3 +37,16 @@ export class CoreError extends Error {
     this.message = message;
   }
 }
+
+type ErrorDataCode = {
+  data?: {
+    code?: string;
+  };
+};
+
+export function getErrorCode(error: unknown) {
+  if (!error || typeof error !== "object") return;
+
+  if ("code" in error && typeof error.code === "string") return error.code;
+  if ("data" in error) return (error as ErrorDataCode).data?.code;
+}
