+++ locales/en/sidebar.json
@@ -22,6 +22,7 @@
   },
   "links": {
     "home": "Home",
+    "cloudblink": "CloudBlink",
     "account": "Account",
     "overview": "Overview",
     "settings": "Settings",
